refactor: reorganize files
