+++ src/components/containers/Header.js
@@ -1,6 +1,6 @@
 import { View } from "react-native-web";
-import Title from "./Title";
-import SubTitle from "./SubTitle";
+import Title from "./domains/headers/Title";
+import SubTitle from "./domains/headers/SubTitle";
 
 
 export default function Header() {
